pyproject.toml
[tool.poetry]
name = "sayge-ai"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "3.8.5"
pytest = "^5.4.1"
pytest-xdist = "^1.31.0"
ipython = "^7.13.0"
ipdb = "^0.13.2"
django = "^3.0.7"
pytest-django = "^3.9.0"
djangorestframework = "^3.11.0"
django-heroku-postgresify = "^0.4"
psycopg2-binary = "^2.8.5"
django-cors-headers = "^3.2.1"
gunicorn = "^20.0.4"
whitenoise = "^5.0.1"
django-extensions = "^2.2.9"
flake8 = "^3.8.1"
sentry-sdk = "0.14.4"
snowflake-connector-python = "^2.2.9"
djoser = "^2.0.3"
pandas = "^1.1.0"
freezegun = "^0.3.15"
django-webpack-loader = "^0.7.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
